+++ src/symbolic/symbolic_choice.ml
@@ -236,6 +236,8 @@ module CoreImpl = struct
 
     val with_thread : (thread -> 'a) -> 'a t
 
+    val lazily : (unit -> 'a) -> 'a t
+
     val set_thread : thread -> unit t
 
     val modify_thread : (thread -> thread) -> unit t
@@ -274,6 +276,8 @@ module CoreImpl = struct
     let with_thread (f : Thread.t -> 'a) : 'a t =
       let x = State.with_state (fun st -> (f st, st)) in
       lift x
+
+    let lazily f = with_thread (fun _th -> f ())
 
     let thread = with_thread Fun.id
 
@@ -393,12 +397,20 @@ module Make (Thread : Thread.S) = struct
       let true_branch =
         let* () = add_pc v in
         let* () = add_breadcrumb 1l in
+        let* () =
+          lazily (fun () ->
+            Stats.event "check true branch reachability" "branches" )
+        in
         let+ () = check_reachability in
         true
       in
       let false_branch =
         let* () = add_pc (Symbolic_value.Bool.not v) in
         let* () = add_breadcrumb 0l in
+        let* () =
+          lazily (fun () ->
+            Stats.event "check true branch reachability" "branches" )
+        in
         let+ () = check_reachability in
         false
       in
@@ -452,7 +464,8 @@ module Make (Thread : Thread.S) = struct
         in
         let this_val_branch =
           let* () = add_breadcrumb i in
-          let+ () = add_pc this_value_cond in
+          let* () = add_pc this_value_cond in
+          let+ () = lazily (fun () -> Stats.event "selected n" "branches") in
           i
         in
         let not_this_val_branch =